requirements.txt
tensorflow>=2.10.0
gradio>=4.0.0
matplotlib>=3.5.0
numpy>=1.21.0
scikit-learn>=1.0.2
pillow>=9.0.0
